calculate measure x ULN ratios and add control; still need to give control functionality
