+++ src/configuration/index.js
@@ -1,0 +1,11 @@
+import settings from './settings';
+import syncSettings from './syncSettings';
+import controlInputs from './controlInputs';
+import syncControlInputs from './syncControlInputs';
+
+export default {
+    settings,
+    syncSettings,
+    controlInputs,
+    syncControlInputs,
+}
