+++ src/wrapper.js
@@ -3,10 +3,9 @@ import './polyfills/array-find.js';
 import './polyfills/moveTo.js';
 import clone from './util/clone.js';
 
-import defaultSettings from './defaultSettings';
-import { syncSettings, syncControlInputs } from './defaultSettings';
-import { createChart, createControls } from 'webcharts';
+import configuration from './configuration/index';
 
+import { createChart, createControls } from 'webcharts';
 import onInit from './callbacks/onInit';
 import onLayout from './callbacks/onLayout';
 import onPreprocess from './callbacks/onPreprocess';
@@ -15,13 +14,14 @@ import onDraw from './callbacks/onDraw';
 import onResize from './callbacks/onResize';
 
 export default function safetyedish(element, settings) {
-    const initial_settings = clone(settings),
-        defaultSettings_clone = clone(defaultSettings),
-        mergedSettings = Object.assign({}, defaultSettings_clone, settings),
-        syncedSettings = syncSettings(mergedSettings),
-        syncedControlInputs = syncControlInputs(syncedSettings),
-        controls = createControls(element, { location: 'top', inputs: syncedControlInputs }),
-        chart = createChart(element, syncedSettings, controls);
+    const initial_settings = clone(settings);
+    const defaultSettings = configuration.settings();
+    const controlInputs = configuration.controlInputs();
+    const mergedSettings = Object.assign({}, defaultSettings, settings);
+    const syncedSettings = configuration.syncSettings(mergedSettings);
+    const syncedControlInputs = configuration.syncControlInputs(controlInputs, syncedSettings);
+    const controls = createControls(element, { location: 'top', inputs: syncedControlInputs });
+    const chart = createChart(element, syncedSettings, controls);
 
     chart.element = element;
     chart.initial_settings = initial_settings;
